include TFXIO module in tests and create and tested APIS
