+++ tfx_bsl/beam/run_inference_record_batch.py
@@ -1,0 +1,57 @@
+# Copyright 2019 Google LLC. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# Lint as: python3
+"""Private API of inference."""
+
+from __future__ import absolute_import
+from __future__ import division
+# Standard __future__ imports
+from __future__ import print_function
+
+import apache_beam as beam
+import tensorflow as tf
+import pyarrow as pa
+from typing import Text, Optional
+from tfx_bsl.beam import run_inference
+from tfx_bsl.public.proto import model_spec_pb2
+from tensorflow_serving.apis import prediction_log_pb2
+
+@beam.ptransform_fn
+@beam.typehints.with_input_types(pa.RecordBatch)
+@beam.typehints.with_output_types(prediction_log_pb2.PredictionLog)
+def RunInferenceOnRecordBatch(  # pylint: disable=invalid-name
+    examples: beam.pvalue.PCollection,
+    inference_spec_type: model_spec_pb2.InferenceSpecType, data_type: Text
+    tensor_adapter_config: Optional[tensor_adapter.TensorAdapterConfig] = None
+) -> beam.pvalue.PCollection:
+  """Run inference with a model.
+
+   There are two types of inference you can perform using this PTransform:
+   1. In-process inference from a SavedModel instance. Used when
+     `saved_model_spec` field is set in `inference_spec_type`.
+   2. Remote inference by using a service endpoint. Used when
+     `ai_platform_prediction_model_spec` field is set in
+     `inference_spec_type`.
+
+  Args:
+    examples: A PCollection containing RecordBatch.
+    inference_spec_type: Model inference endpoint.
+
+  Returns:
+    A PCollection containing prediction logs.
+  """
+
+  return (
+      examples | 'RunInferenceImpl' >> run_inference.RunInferenceImpl(
+                        inference_spec_type, data_type, tensor_adapter_config))
