+++ tfx_bsl/public/beam/run_inference.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # Lint as: python3
-"""Publich API of batch inference."""
+"""Public API of batch inference."""
 
 from __future__ import absolute_import
 from __future__ import division
@@ -28,7 +28,6 @@ from tfx_bsl.tfxio import tensor_adapter
 from tfx_bsl.tfxio import tf_example_record
 from tfx_bsl.tfxio import tf_sequence_example_record
 from tfx_bsl.beam import run_inference
-from tfx_bsl.beam import run_inference_arrow
 from tfx_bsl.public.proto import model_spec_pb2
 from tensorflow_serving.apis import prediction_log_pb2
 from tensorflow_metadata.proto.v0 import schema_pb2
@@ -71,6 +70,7 @@ def RunInference(  # pylint: disable=invalid-name
     schema=schema,
     raw_record_column_name=_RECORDBATCH_COLUMN)
 
+  tensor_adapter_config = None
   if schema:
     tfxio = test_util.InMemoryTFExampleRecord(
       schema=schema, raw_record_column_name=_RECORDBATCH_COLUMN)
@@ -79,9 +79,9 @@ def RunInference(  # pylint: disable=invalid-name
       tensor_representations=tfxio.TensorRepresentations())
 
   return (examples
-          | 'ParseExamples' >> beam.Map(tf.train.Example.SerializeToString)
+          | 'ParseExamples' >> beam.Map(lambda example: example.SerializeToString())
           | 'ConvertToRecordBatch' >> converter.BeamSource()
-          | 'RunInferenceImpl' >> run_inference_arrow.RunInferenceImpl(
+          | 'RunInferenceImpl' >> run_inference.RunInferenceImpl(
                   inference_spec_type, data_type,
                   tensor_adapter_config=tensor_adapter_config))
 
@@ -120,6 +120,7 @@ def RunInferenceOnSequenceExamples(  # pylint: disable=invalid-name
     schema=schema,
     raw_record_column_name=_RECORDBATCH_COLUMN)
 
+  tensor_adapter_config = None
   if schema:
     tfxio = test_util.InMemoryTFExampleRecord(
       schema=schema, raw_record_column_name=_RECORDBATCH_COLUMN)
@@ -128,38 +129,8 @@ def RunInferenceOnSequenceExamples(  # pylint: disable=invalid-name
       tensor_representations=tfxio.TensorRepresentations())
 
   return (examples
-          | 'ParseExamples' >> beam.Map(tf.train.Example.SerializeToString)
+          | 'ParseExamples' >> beam.Map(lambda example: example.SerializeToString())
           | 'ConvertToRecordBatch' >> converter.BeamSource()
-          | 'RunInferenceImpl' >> run_inference_arrow.RunInferenceImpl(
+          | 'RunInferenceImpl' >> run_inference.RunInferenceImpl(
                   inference_spec_type, data_type,
                   tensor_adapter_config=tensor_adapter_config))
-
-
-@beam.ptransform_fn
-@beam.typehints.with_input_types(pa.RecordBatch)
-@beam.typehints.with_output_types(prediction_log_pb2.PredictionLog)
-def RunInferenceOnRecordBatch(  # pylint: disable=invalid-name
-    examples: beam.pvalue.PCollection,
-    inference_spec_type: model_spec_pb2.InferenceSpecType,
-    tensor_adapter_config: Optional[tensor_adapter.TensorAdapterConfig] = None
-) -> beam.pvalue.PCollection:
-  """Run inference with a model.
-
-   There are two types of inference you can perform using this PTransform:
-   1. In-process inference from a SavedModel instance. Used when
-     `saved_model_spec` field is set in `inference_spec_type`.
-   2. Remote inference by using a service endpoint. Used when
-     `ai_platform_prediction_model_spec` field is set in
-     `inference_spec_type`.
-
-  Args:
-    examples: A PCollection containing RecordBatch.
-    inference_spec_type: Model inference endpoint.
-
-  Returns:
-    A PCollection containing prediction logs.
-  """
-
-  return (
-      examples | 'RunInferenceImpl' >> run_inference_arrow.RunInferenceImpl(
-                        inference_spec_type, tensor_adapter_config))
